feat: Implement SessionCard component and integrate with SessionsLists for improved session display
